DICOM-Annotator-server/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
numpy==1.19.3
Pillow==8.0.1
pydicom==2.1.1
pypng==0.0.20
pytz==2020.4
sqlparse==0.4.1
